contract_analyzer.py

import os
import re
import json
import argparse
from enum import Enum
from dataclasses import dataclass, field
from typing import List, Dict, Optional, Literal
import docx
from pdfminer.high_level import extract_text


# Type definitions
class RiskLevel(Enum):
    LOW = "low"
    MEDIUM = "medium" 
    HIGH = "high"


@dataclass
class Clause:
    id: str
    title: str
    content: str
    type: str
    risk: RiskLevel


@dataclass
class Contract:
    id: str
    title: str
    parties: List[str]
    clauses: List[Clause] = field(default_factory=list)


@dataclass
class EditSuggestion:
    id: str
    clause_id: str
    original: str
    suggested: str
    reasoning: str
    impact: str


@dataclass
class CounterPartyArgument:
    id: str
    clause_id: str
    argument: str
    strength: Literal["weak", "moderate", "strong"]


@dataclass
class LegalReference:
    id: str
    clause_id: str
    source: str
    citation: str
    relevance: str
    url: Optional[str] = None


# Document parsing
def parse_document(file_path: str) -> str:
    """Extract text from a document file (PDF or DOCX)."""
    _, file_extension = os.path.splitext(file_path)
    
    if file_extension.lower() == '.pdf':
        return extract_text(file_path)
    elif file_extension.lower() == '.docx':
        doc = docx.Document(file_path)
        return '\n'.join([paragraph.text for paragraph in doc.paragraphs])
    else:
        raise ValueError(f"Unsupported file format: {file_extension}")


def identify_clauses(document_text: str) -> List[dict]:
    """
    Identify clauses in a document text.
    This is a simplified implementation and would need more advanced NLP in a real application.
    """
    # Simple pattern matching for clause identification
    # In a real app, this would use more sophisticated NLP techniques
    clause_pattern = r'(?i)(\d+\.?\s*[A-Z][^\.]+\.)' 
    matches = re.finditer(clause_pattern, document_text)
    
    clauses = []
    for i, match in enumerate(matches):
        clauses.append({
            "id": f"clause_{i+1}",
            "title": f"Clause {i+1}",
            "content": match.group(0),
            "type": "general",  # Would need more advanced analysis to determine type
            "risk": RiskLevel.MEDIUM.value  # Default risk level
        })
    
    return clauses


def analyze_risk(clause_text: str) -> RiskLevel:
    """
    Analyze the risk level of a clause.
    This is a placeholder implementation.
    """
    # For demonstration purposes, we'll use simple keyword-based analysis
    high_risk_terms = ["terminate", "liability", "indemnity", "penalty", "breach"]
    medium_risk_terms = ["change", "modify", "amend", "extend", "renew"]
    
    text_lower = clause_text.lower()
    
    for term in high_risk_terms:
        if term in text_lower:
            return RiskLevel.HIGH
    
    for term in medium_risk_terms:
        if term in text_lower:
            return RiskLevel.MEDIUM
    
    return RiskLevel.LOW


def generate_edit_suggestions(clause: Clause) -> List[EditSuggestion]:
    """
    Generate edit suggestions for a clause.
    This is a placeholder implementation.
    """
    # In a real application, this would use NLP or a language model
    if clause.risk == RiskLevel.HIGH.value:
        return [
            EditSuggestion(
                id=f"suggestion_{clause.id}_1",
                clause_id=clause.id,
                original=clause.content,
                suggested=f"Consider revising: {clause.content}",
                reasoning="High risk clause that may expose to liability",
                impact="Reduces legal exposure"
            )
        ]
    return []


def simulate_counterparty_arguments(clause: Clause) -> List[CounterPartyArgument]:
    """
    Simulate possible counterparty arguments.
    This is a placeholder implementation.
    """
    # In a real application, this would use more sophisticated analysis
    if clause.risk == RiskLevel.HIGH.value:
        return [
            CounterPartyArgument(
                id=f"argument_{clause.id}_1",
                clause_id=clause.id,
                argument=f"This clause is industry standard and shouldn't be modified.",
                strength="strong"
            )
        ]
    elif clause.risk == RiskLevel.MEDIUM.value:
        return [
            CounterPartyArgument(
                id=f"argument_{clause.id}_1",
                clause_id=clause.id,
                argument=f"We can discuss minor modifications but the core terms must remain.",
                strength="moderate"
            )
        ]
    return []


def find_legal_references(clause: Clause) -> List[LegalReference]:
    """
    Find relevant legal references for a clause.
    This is a placeholder implementation.
    """
    # In a real application, this would query a legal database
    return [
        LegalReference(
            id=f"reference_{clause.id}_1",
            clause_id=clause.id,
            source="General Contract Law",
            citation="Section 2-207 of the UCC",
            relevance="Provides context for contract formation and modification",
            url="https://www.law.cornell.edu/ucc/2/2-207"
        )
    ]


def analyze_contract(contract: Contract) -> Dict:
    """Analyze a contract and return the results."""
    results = {
        "contract_id": contract.id,
        "contract_title": contract.title,
        "analysis": []
    }
    
    for clause in contract.clauses:
        # Update risk assessment
        risk_level = analyze_risk(clause.content)
        clause.risk = risk_level.value
        
        # Generate analysis
        clause_analysis = {
            "clause_id": clause.id,
            "clause_title": clause.title,
            "risk_level": clause.risk,
            "edit_suggestions": [suggestion.__dict__ for suggestion in generate_edit_suggestions(clause)],
            "counterparty_arguments": [arg.__dict__ for arg in simulate_counterparty_arguments(clause)],
            "legal_references": [ref.__dict__ for ref in find_legal_references(clause)]
        }
        
        results["analysis"].append(clause_analysis)
    
    return results


def process_document(file_path: str) -> Dict:
    """Process a document file and return analysis results."""
    # Parse document
    document_text = parse_document(file_path)
    
    # Identify clauses
    clauses_data = identify_clauses(document_text)
    clauses = [Clause(**clause_data) for clause_data in clauses_data]
    
    # Create contract object
    contract = Contract(
        id="contract_1",
        title=os.path.basename(file_path),
        parties=["Your Company", "Counterparty"],  # This would need to be extracted in a real app
        clauses=clauses
    )
    
    # Analyze contract
    results = analyze_contract(contract)
    
    return results


def main():
    parser = argparse.ArgumentParser(description='Analyze contract documents')
    parser.add_argument('file_path', help='Path to the contract document (PDF or DOCX)')
    parser.add_argument('--output', help='Output file path for analysis results (JSON)')
    
    args = parser.parse_args()
    
    try:
        results = process_document(args.file_path)
        
        if args.output:
            with open(args.output, 'w') as f:
                json.dump(results, f, indent=2)
            print(f"Analysis results saved to {args.output}")
        else:
            print(json.dumps(results, indent=2))
            
    except Exception as e:
        print(f"Error processing document: {e}")


if __name__ == "__main__":
    main()
