requirements.txt

pdfminer.six==20221105
python-docx==0.8.11
